+++ packages/react/lib/DropdownToggle/index.tsx
@@ -1,16 +1,17 @@
 import {
   type RefObject,
-  type ReactElement,
-  type ComponentPropsWithoutRef,
   cloneElement,
   useImperativeHandle,
   useRef,
+  use,
 } from 'react';
 import { classNames } from '@junipero/core';
 
 import type {
   JuniperoInnerRef,
   JuniperoRef,
+  ReactElt,
+  ReactLazy,
   SpecialComponentPropsWithRef,
 } from '../types';
 import { useDropdown } from '../hooks';
@@ -34,25 +35,27 @@ const DropdownToggle = ({
     isJunipero: true,
   }));
 
-  const child: ReactElement<
-    ComponentPropsWithoutRef<any>
-  > = typeof children !== 'string' && Array.isArray(children)
-    ? children[0] : children;
+  const child: ReactElt | ReactLazy =
+    typeof children !== 'string' && Array.isArray(children)
+      ? children[0] : children;
 
-  return cloneElement(child, {
-    className: classNames(
-      child.props?.className, 'dropdown-toggle', { opened }
-    ),
-    ref: (r: JuniperoRef | JuniperoInnerRef) => {
-      innerRef.current = (r as JuniperoRef)?.isJunipero
-        ? (r as JuniperoRef).innerRef.current : r;
-      refs.setReference(
-        (r as JuniperoRef)?.isJunipero
-          ? (r as JuniperoRef).innerRef.current : r
-      );
-    },
-    ...getReferenceProps({ onClick: child.props?.onClick }),
-  });
+  return cloneElement(
+    (child as ReactLazy).$$typeof === Symbol.for('react.lazy')
+      ? use<ReactElt>((child as ReactLazy)._payload) : child as ReactElt,
+    {
+      className: classNames(
+        (child as ReactElt).props?.className, 'dropdown-toggle', { opened }
+      ),
+      ref: (r: JuniperoRef | JuniperoInnerRef) => {
+        innerRef.current = (r as JuniperoRef)?.isJunipero
+          ? (r as JuniperoRef).innerRef.current : r;
+        refs.setReference(
+          (r as JuniperoRef)?.isJunipero
+            ? (r as JuniperoRef).innerRef.current : r
+        );
+      },
+      ...getReferenceProps({ onClick: (child as ReactElt).props?.onClick }),
+    });
 };
 
 DropdownToggle.displayName = 'DropdownToggle';